+++ features/foundation-info/src/main/java/com/example/foundation_info/domain/repository/FoundationRepository.kt
@@ -4,5 +4,5 @@ import com.example.foundation_info.domain.entity.FoundationEntity
 
 interface FoundationRepository {
 
-    suspend fun getFoundation(query: Int): FoundationEntity
+    suspend fun getFoundation(query: Long): FoundationEntity
 }
